# api/openapi.yaml
openapi: 3.0.0
info:
  description: The API for the simple containerisation solution Podinate.
  title: Podinate API
  version: 0.0.1
servers:
- description: Production server
  url: https://api.podinate.com/v0
- description: Development Server
  url: http://localhost:3001/v0
security:
  - APIKeyAuth: []
...
paths:
  /account:
    get:
      summary: Get information about the current account.
      description: Accounts are logically separate, distinct containers for different customers or teams. For example you might have an account for your marketing and product teams, or you might have an account for each client to separate their workloads. Accounts can have their own security settings applied.
      tags:
      - Account
      parameters:
        - $ref: '#/components/parameters/account'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
          description: Information about the given account
        "404":
          $ref: "#/components/responses/404Error"
        "500":
          $ref: "#/components/responses/500Error"
        "409":
          $ref: "#/components/responses/409Error"
    post:
      summary: Create a new account
      description: Create a new logically separate account, for example for a given client or team.
      tags:
      - Account
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
      responses:
        "201":
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        "400":
          $ref: "#/components/responses/400Error"
        "500":
          $ref: "#/components/responses/500Error"
    put:
      summary: Update an existing account
      description: Update an existing account, for example to change the display name
      tags:
      - Account
      parameters:
      - $ref: '#/components/parameters/account'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
      responses:
        "200":
          description: Account updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        "400":
          $ref: "#/components/responses/400Error"
        "500":
          $ref: "#/components/responses/500Error"
          description: The server was not able to update your account
    delete:
      summary: Delete the account and all associated resources!
      description: Deletes the account and all associated resources! This is irreversible and may be used for example when a client project has finished or a migration is complete. 
      tags:
      - Account
      parameters:
      - $ref: '#/components/parameters/account'
      responses:
        "202":
          description: Podinate receieved your request to delete the account - it will be processed in the background. 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        "400":
          $ref: "#/components/responses/400Error"
        "404":
          $ref: "#/components/responses/404Error"
        "500":
          $ref: "#/components/responses/500Error"
...
  /project/{project_id}/shared_volumes:
    get:
      summary: Get a list of shared volumes for a given project
      description: "Returns a list of shared volumes for a given project. Shared volumes are storage volumes that can be shared by multiple pods to store data."
      tags:
      - SharedVolume
      parameters:
      # Pagination parameters
      - in: query
        name: page
        schema:
          type: integer
          default: 0
        description: The page number to return (starts at 0)
      - in: query
        name: limit
        schema:
          type: integer
          default: 20
        description: The amount of items to return per page
          # example: 10

      - explode: false
        in: path
        name: project_id
        required: true
        schema:
          description: The project to look at
          example: hello-world
          type: string
        style: simple
      - $ref: '#/components/parameters/account'
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SharedVolume'
                  total:
                    type: integer
                    description: The total number of shared volumes
                    example: 1
                  page:
                    type: integer
                    description: The current page number
                    example: 0
                  limit:
                    type: integer
                    description: The number of items per page
                    example: 20
                required:
                - total
                - page
                - limit
          description: A JSON array of shared volumes
        "400":
          $ref: "#/components/responses/400Error"
        "404":
          $ref: "#/components/responses/404Error"
        "500":
          $ref: "#/components/responses/500Error"
        # Add conflict
    post:
      summary: Create a new shared volume
      description: Creating a new shared volume
      tags:
      - SharedVolume
      parameters:
      - explode: false
        in: path
        name: project_id
        required: true
        schema:
          description: The project to look at
          example: hello-world
          type: string
        style: simple
      - $ref: '#/components/parameters/account'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SharedVolume'
        description: A JSON object containing the information needed to create a new
          shared volume
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedVolume'
          description: The shared volume was created successfully
        "400":
          $ref: "#/components/responses/400Error"
        "500":
          $ref: "#/components/responses/500Error"
  /project/{project_id}/shared_volumes/{volume_id}:
    get:
      summary: Get a shared volume by ID
      description: Get a shared volume by ID
      tags:
      - SharedVolume
      parameters:
      - explode: false
        in: path
        name: project_id
        required: true
        schema:
          description: The project to look at
          example: hello-world
          type: string
        style: simple
      - explode: false
        in: path
        name: volume_id
        required: true
        schema:
          description: The shared volume to look at
          example: hello-world
          type: string
        style: simple
        x-speakeasy-match: id
      - $ref: '#/components/parameters/account'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedVolume'
          description: An array of shared volumes
        "400":
          $ref: "#/components/responses/400Error"
        "404":
          $ref: "#/components/responses/404Error"
        "500":
          $ref: "#/components/responses/500Error"
    put:
      summary: Update a shared volume's spec
      description: Update a shared volume
      tags:
      - SharedVolume
      parameters:
      - explode: false
        in: path
        name: project_id
        required: true
        schema:
          description: The project to look at
          example: hello-world
          type: string
        style: simple
      - explode: false
        in: path
        name: volume_id
        required: true
        schema:
          description: The shared volume to look at
          example: hello-world
          type: string
        style: simple
        x-speakeasy-match: id
      - $ref: '#/components/parameters/account'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SharedVolume'
        description: A JSON object containing the information needed to update a shared volume
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedVolume'
          description: The shared volume was updated successfully
        "400":
          $ref: "#/components/responses/400Error"
    delete:
      summary: Delete a shared volume
      description: Delete a shared volume
      tags:
      - SharedVolume
      parameters:
      - explode: false
        in: path
        name: project_id
        required: true
        schema:
          description: The project to look at
          example: hello-world
          type: string
        style: simple
      - explode: false
        in: path
        name: volume_id
        required: true
        schema:
          description: The shared volume to look at
          example: hello-world
          type: string
        style: simple
        x-speakeasy-match: id
      - $ref: '#/components/parameters/account'
      responses:
        "202":
          description: We received your request to delete the shared volume, it will be processed in the background. 
        "400":
          $ref: "#/components/responses/400Error"
        "404":
          $ref: "#/components/responses/404Error"
        "500":
          $ref: "#/components/responses/500Error"
...
components:
  securitySchemes:
    APIKeyAuth:
      type: apiKey
      in: header
      name: Authorization
  parameters:
    account:
      description: The account to use for the request
      example: my-account
      explode: false
      in: header
      name: account
      required: true
      schema:
        type: string
      style: simple
...
  responses:
    "400Error":
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: "Request issued incorrectly, for example missing parameters or\
        \ wrong endpoint"
    "404Error":
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: "The requested resource was not found"
    "500Error":
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: An internal server error
    "409Error":
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: "The request could not be completed due to a conflict with the\
        \ current state of the resource"
  schemas:
    Account:
      example:
        id: "blog-prod"
        name: "Company Blog - Prod"
      properties:
        id:
          description: The short name (slug/url) of the account. The account ID is globally unique and cannot be changed.
          example: "blog-prod"
          type: string 
        name:
          description: The human readable name of the account, used for display purposes.
          example: "Tim's Widgets Company Blog - Prod"
          type: string
        resource_id:
          description: The global Resource ID of the account
          example: account:podinate-blog-prod
          type: string
          x-speakeasy-param-readonly: true
...
    SharedVolume:
      description: A shared volume is a storage volume that can be shared by multiple pods to store data. 
      properties:
        id:
          description: The short name (slug/url) of the shared volume
          example: blog-files
          type: string
        name:
          description: The name of the shared volume
          example: "Blog Files"
          type: string
        size:
          description: The size of the shared volume in GB
          example: 10
          type: integer
        class:
          description: The class of the shared volume, for example "standard" or "premium"
          example: standard
          type: string
        resource_id:
          description: The global Resource ID of the shared volume
          example: account:podinate/project:hello-world/shared_volume:blog-files
          type: string
          x-speakeasy-param-readonly: true
      required:
        - id
        - size
      type: object
    "Error":
      properties:
        code:
          default: 400
          description: The http code of the error message (400)
          example: 400
          type: number
        message:
          description: The human-readable description of what went wrong
          example: Missing required parameter 'account'
          type: string
      type: object